requirements.txt
faker>=26.0.0,<27.0.0
selenium>=4.25.0,<5.0.0
webdriver-manager>=4.0.2,<5.0.0
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
